FEAT: 홈 프로필 파트 조회 성공 후 매핑 (#88)
